Added Icon for our project easter egg:
Now starring: The Annoying Dick Preston Gravey and The Settlements That Needs Your help.

At each route a POI will be placed with the icon of Preston, after activating a message is displayed with: I heard that another settlement needs our help, I will mark it on your map.

Then a new POI is added with the Icon of a settlement.

Maybe the new idea will be to add a quest line, but I think we can't do beacuase of the copyright issues